Split up SaltAPIBuilder file
