+++ src/main/java/com/waytta/SaltAPIBuilder.java
@@ -167,7 +167,7 @@ public class SaltAPIBuilder extends Builder implements SimpleBuildStep {
         boolean myBlockBuild = getBlockbuild();
         boolean jobSuccess = true;
 
-        StandardUsernamePasswordCredentials credential = getCredentialById(getCredentialsId());
+        StandardUsernamePasswordCredentials credential = Utils.getCredentialById(getCredentialsId());
         if (credential == null) {
             listener.error("Invalid credentials");
             return false;
@@ -175,7 +175,7 @@ public class SaltAPIBuilder extends Builder implements SimpleBuildStep {
 
         // Setup connection for auth
         String token = new String();
-        JSONObject auth = createAuthArray(credential);
+        JSONObject auth = Utils.createAuthArray(credential, authtype);
         JSONObject httpResponse = new JSONObject();
         JSONArray returnArray = new JSONArray();
 
@@ -193,8 +193,10 @@ public class SaltAPIBuilder extends Builder implements SimpleBuildStep {
 
         // Access different salt-api endpoints depending on function
         if (myBlockBuild) {
+        	int jobPollTime = getJobPollTime();
+            int minionTimeout = getDescriptor().getTimeoutTime();
             // poll /minion for response
-        	jobSuccess = runBlockingBuild(returnArray, myservername, token, saltFunc, listener);
+        	jobSuccess = Builds.runBlockingBuild(returnArray, myservername, token, saltFunc, listener, jobPollTime, minionTimeout);
         } else if (myClientInterface.equals("hook")) {
         	// publish event to salt event bus to /hook
         	String myTag = Utils.paramorize(build, listener, getTag());
@@ -251,237 +253,49 @@ public class SaltAPIBuilder extends Builder implements SimpleBuildStep {
         // Results now printed. Return success condition 
         return jobSuccess;
     }
-
-    private JSONObject createAuthArray(StandardUsernamePasswordCredentials credential) {
-        JSONObject auth = new JSONObject();
-        auth.put("username", credential.getUsername());
-        auth.put("password", credential.getPassword().getPlainText());
-        auth.put("eauth", authtype);
-
-        return auth;
-    }
-
-    private JSONObject prepareSaltFunction(Run build, TaskListener listener, String myClientInterface,
-                                           String mytarget, String myfunction, String myarguments 
-                                           ) throws IOException, InterruptedException {
-        JSONObject saltFunc = new JSONObject();
-        saltFunc.put("client", myClientInterface);
-
-        switch (myClientInterface) {
-        case "local_batch":
-        	String mybatch = Utils.paramorize(build, listener, getBatchSize());
-            saltFunc.put("batch", mybatch);
-            listener.getLogger().println("Running in batch mode. Batch size: " + mybatch);
-            break;
-        case "runner":
-        	saltFunc.put("mods", getMods());
-            String myPillarvalue = Utils.paramorize(build, listener, getPillarvalue());
-            if ( myPillarvalue.length() > 0) {
-            	JSONObject jPillar = new JSONObject();
-            	// If value was already a jsonobject, treat it as such
-            	JSON runPillarValue = JSONSerializer.toJSON(myPillarvalue);
-            	saltFunc.put("pillar", runPillarValue);
-            }
-            break;
-        case "local_subset":
-        	String mySubset = Utils.paramorize(build, listener, getSubset());
-        	saltFunc.put("sub", Integer.parseInt(mySubset));
-            listener.getLogger().println("Running in subset mode. Subset size: " + mySubset);
-            break;
-        case "hook":
-        	// Posting to /hook url should only contain object to be posted
-        	String myPost = Utils.paramorize(build, listener, getPost());
-        	saltFunc = JSONObject.fromObject(myPost);
-        	return saltFunc;
-        }
-
-        saltFunc.put("tgt", mytarget);
-        saltFunc.put("expr_form", getTargettype());
-        saltFunc.put("fun", myfunction);
-        addArgumentsToSaltFunction(myarguments, saltFunc);
-
-        return saltFunc;
-    }
-
-
-    private void addArgumentsToSaltFunction(String myarguments, JSONObject saltFunc) {
-        if (myarguments.length() > 0) {
-            JSONObject fullKwJSON = new JSONObject();
-
-            // spit on space separated not inside of single and double quotes
-            String[] argItems = myarguments.split("\\s+(?=(?:[^'\"]|'[^']*'|\"[^\"]*\")*$)");
-
-            for (String arg : argItems) {
-                // remove spaces at beginning or end
-                arg = arg.replaceAll("^\\s+|\\s+$", "");
-                // if string wrapped in quotes, remove them since adding to list
-                // re-quotes
-                arg = arg.replaceAll("(^')|(^\")|('$)|(\"$)", "");
-                if (arg.contains("=")) {
-                    String[] kwString = arg.split("=");
-                    if (kwString.length > 2) {
-                        // kwarg contained more than one =. Let's put the string
-                        // back together
-                        String kwFull = new String();
-                        for (String kwItem : kwString) {
-                            // Ignore the first item as it will remain the key
-                            if (kwItem == kwString[0]) {
-                                continue;
-                            }
-                            // add the second item
-                            if (kwItem == kwString[1]) {
-                            	// again remove any wrapping quotes
-                                kwItem = kwItem.replaceAll("(^')|(^\")|('$)|(\"$)", "");
-                                kwFull += kwItem;
-                                continue;
-                            }
-                            // add all other items with an = to rejoin
-                            kwFull += "=" + kwItem;
-                        }
-                        fullKwJSON.put(kwString[0], kwFull);
-                    } else {
-                    	// again remove any wrapping quotes
-                        kwString[1] = kwString[1].replaceAll("(^')|(^\")|('$)|(\"$)", "");
-        	        	try {
-        	                // If value was already a jsonobject, treat it as such
-        	                JSON kwJSON = JSONSerializer.toJSON(kwString[1]);
-        	                // Rejoin key and json object
-        	                fullKwJSON.element(kwString[0], kwJSON);
-        	            } catch (JSONException e) {
-        	                // Otherwise it must have been a string
-                            fullKwJSON.put(kwString[0], kwString[1]);
-        	            }
-                    }
-                } else {
-                	// Add any args to json message
-                	saltFunc.accumulate("arg", arg);
-                }
-            }
-            // Now that loops have completed, add kwarg object
-            saltFunc.element("kwarg", fullKwJSON);
-        }
-    }
     
-    private boolean runBlockingBuild(JSONArray returnArray, String myservername, 
-    		String token, JSONObject saltFunc, TaskListener listener) throws InterruptedException {
-    	JSONObject httpResponse = new JSONObject();
-    	String jid = new String();
-    	// Send request to /minion url. This will give back a jid which we
-    	// will need to poll and lookup for completion
-    	httpResponse = Utils.getJSON(myservername + "/minions", saltFunc, token);
-    	try {
-    		returnArray = httpResponse.getJSONArray("return");
-    		for (Object o : returnArray) {
-    			JSONObject line = (JSONObject) o;
-    			jid = line.getString("jid");
-    		}
-    		// Print out success
-    		listener.getLogger().println("Running jid: " + jid);
-    	} catch (Exception e) {
-    		listener.error(httpResponse.toString(2));
-    		return false;
-    	}
+    private JSONObject prepareSaltFunction(Run build, TaskListener listener, String myClientInterface, String mytarget,
+			String myfunction, String myarguments) throws IOException, InterruptedException {
+		JSONObject saltFunc = new JSONObject();
+		saltFunc.put("client", myClientInterface);
 
-    	// Request successfully sent. Now use jid to check if job complete
-    	int numMinions = 0;
-    	int numMinionsDone = 0;
-    	JSONArray minionsArray = new JSONArray();
-    	httpResponse = Utils.getJSON(myservername + "/jobs/" + jid, null, token);
-    	try {
-    		// info array will tell us how many minions were targeted
-    		returnArray = httpResponse.getJSONArray("info");
-    		for (Object o : returnArray) {
-    			JSONObject line = (JSONObject) o;
-    			minionsArray = line.getJSONArray("Minions");
-    			// Check the info[Minions[]] array to see how many nodes we
-    			// expect to hear back from
-    			numMinions = minionsArray.size();
-    			listener.getLogger().println("Waiting for " + numMinions + " minions");
-    		}
-    		returnArray = httpResponse.getJSONArray("return");
-    		// Check the return[] array to see how many minions have
-    		// responded
-    		if (!returnArray.getJSONObject(0).names().isEmpty()) {
-    			numMinionsDone = returnArray.getJSONObject(0).names().size();
-    		} else {
-    			numMinionsDone = 0;
-    		}
-    		listener.getLogger().println(numMinionsDone + " minions are done");
-    	} catch (Exception e) {
-    		listener.error(httpResponse.toString(2));
-    		return false;
-    	}
+		switch (myClientInterface) {
+		case "local_batch":
+			String mybatch = Utils.paramorize(build, listener, getBatchSize());
+			saltFunc.put("batch", mybatch);
+			listener.getLogger().println("Running in batch mode. Batch size: " + mybatch);
+			break;
+		case "runner":
+			saltFunc.put("mods", getMods());
+			String myPillarvalue = Utils.paramorize(build, listener, getPillarvalue());
+			if (myPillarvalue.length() > 0) {
+				JSONObject jPillar = new JSONObject();
+				// If value was already a jsonobject, treat it as such
+				JSON runPillarValue = JSONSerializer.toJSON(myPillarvalue);
+				saltFunc.put("pillar", runPillarValue);
+			}
+			break;
+		case "local_subset":
+			String mySubset = Utils.paramorize(build, listener, getSubset());
+			saltFunc.put("sub", Integer.parseInt(mySubset));
+			listener.getLogger().println("Running in subset mode. Subset size: " + mySubset);
+			break;
+		case "hook":
+			// Posting to /hook url should only contain object to be posted
+			String myPost = Utils.paramorize(build, listener, getPost());
+			saltFunc = JSONObject.fromObject(myPost);
+			return saltFunc;
+		}
 
-    	// Now that we know how many minions have responded, and how many we
-    	// are waiting on. Let's see more have finished
-    	if (numMinionsDone < numMinions) {
-    		// Don't print annoying messages unless we really are waiting for
-    		// more minions to return
-    		listener.getLogger().println(
-    				"Will check status every " + getJobPollTime() + " seconds...");
-    	}
-    	while (numMinionsDone < numMinions) {
-    		try {
-    			Thread.sleep(getJobPollTime() * 1000);
-    		} catch (InterruptedException ex) {
-    			Thread.currentThread().interrupt();
-    			// Allow user to cancel job in jenkins interface
-    			throw new InterruptedException();
-    		}
-    		Integer oldMinionsDone = numMinionsDone;
-    		httpResponse = Utils.getJSON(servername + "/jobs/" + jid, null, token);
-    		returnArray = httpResponse.getJSONArray("return");
-    		numMinionsDone = returnArray.getJSONObject(0).names().size();
-            Integer minionTimeout = getDescriptor().getTimeoutTime();
-    		if (numMinionsDone > oldMinionsDone && numMinionsDone < numMinions) {
-    			// Some minions returned, but not all
-    			// Give them minionTimeout to all return or fail build
-    			try {
-    				listener.getLogger().println(
-    						"Some minions returned. Waiting " + minionTimeout + " seconds");
-    				Thread.sleep(minionTimeout * 1000);
-    			} catch (InterruptedException ex) {
-    				Thread.currentThread().interrupt();
-    				// Allow user to cancel job in jenkins interface
-    				throw new InterruptedException();
-    			}
-    			httpResponse = Utils.getJSON(servername + "/jobs/" + jid, null, token);
-    			returnArray = httpResponse.getJSONArray("return");
-    			numMinionsDone = returnArray.getJSONObject(0).names().size();
-    			if (numMinionsDone < numMinions) {
-    				JSONArray respondedMinions = returnArray.getJSONObject(0).names();
-    				for (int i = 0; i < respondedMinions.size(); ++i) {
-    					minionsArray.discard(respondedMinions.get(i));
-    				}
-    				listener.error(
-    						"Minions timed out:\n" + minionsArray.toString() + "\n\n");
-    				return false;
-    			}
-    		}
-    	}
-    	return true;
-    }
-    
-    StandardUsernamePasswordCredentials getCredentialById(String credentialId) {
-        List<StandardUsernamePasswordCredentials> credentials = getCredentials(Jenkins.getInstance());
-        for (StandardUsernamePasswordCredentials credential : credentials) {
-            if (credential.getId().equals(credentialId)) {
-                return credential;
-            }
-        }
-        return null;
-    }
+		saltFunc.put("tgt", mytarget);
+		saltFunc.put("expr_form", getTargettype());
+		saltFunc.put("fun", myfunction);
+		Builds.addArgumentsToSaltFunction(myarguments, saltFunc);
 
-    static List<StandardUsernamePasswordCredentials> getCredentials(Jenkins context) {
-        List<DomainRequirement> requirements = URIRequirementBuilder.create().build();
-        List<StandardUsernamePasswordCredentials> credentials = CredentialsProvider.lookupCredentials(StandardUsernamePasswordCredentials.class, context, ACL.SYSTEM, requirements);
-        return credentials;
-    }
+		return saltFunc;
+	}
     
 
-    // Overridden for better type safety.
-    // If your plugin doesn't really define any property on Descriptor,
-    // you don't have to do this.
     @Override
     public DescriptorImpl getDescriptor() {
         return (DescriptorImpl) super.getDescriptor();
@@ -531,7 +345,7 @@ public class SaltAPIBuilder extends Builder implements SimpleBuildStep {
                 @QueryParameter String credentialsId,
                 @QueryParameter String authtype) {
             StandardUsernamePasswordCredentials usedCredential = null;
-            List<StandardUsernamePasswordCredentials> credentials = getCredentials(Jenkins.getInstance());
+            List<StandardUsernamePasswordCredentials> credentials = Utils.getCredentials(Jenkins.getInstance());
             for (StandardUsernamePasswordCredentials credential : credentials) {
                 if (credential.getId().equals(credentialsId)) {
                     usedCredential = credential;
@@ -565,7 +379,7 @@ public class SaltAPIBuilder extends Builder implements SimpleBuildStep {
             StandardListBoxModel result = new StandardListBoxModel();
             result.withEmptySelection();
             result.withMatching(CredentialsMatchers.instanceOf(StandardUsernamePasswordCredentials.class),
-                    getCredentials(context));
+                    Utils.getCredentials(context));
             return result;
         }
 
@@ -601,7 +415,7 @@ public class SaltAPIBuilder extends Builder implements SimpleBuildStep {
             }
 
             StandardUsernamePasswordCredentials usedCredential = null;
-            List<StandardUsernamePasswordCredentials> credentials = getCredentials(Jenkins.getInstance());
+            List<StandardUsernamePasswordCredentials> credentials = Utils.getCredentials(Jenkins.getInstance());
             for (StandardUsernamePasswordCredentials credential : credentials) {
                 if (credential.getId().equals(value)) {
                     usedCredential = credential;
